feat: implementa CRUD completo para Author Service
- Agrega métodos findAll, save, edit y delete
- Implementa soft delete para autores
- Manejo consistente de errores con orElseThrow

+++ src/main/java/konoec/bookhub/domain/bookManagement/format/FormatRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FormatRepository extends JpaRepository<Format, Integer> {
+public interface FormatRepository extends JpaRepository<Format, Long> {
 }
